+++ ChatApp/etc/Todo
@@ -53,3 +53,4 @@ Fix unread badge
 move profile photo to bottom message
 dock your messages to the side
 make it so when you hover it shows the timestamp and read of that message your hovering on 
+Create MEssage Animation for typing
